Removed Jetty from the Repo
I removed Jetty from the repo because it was a lot of code that skewed our repository towards HTML, instead of the JAva that we've been writing. I kept the 'webapps' folder from it in the repo though, so getting the servlet running should be as easy as copying the contents of webapps into the webapps folder on a local jetty container instance.